Moving PikaConsumer into pika module
